[splitormergeobjects]: Moved major part of the run funciton to library Also moved two helper functions out of FE module. 1. The `filter_using_image()` function to library/module. Might move this to some other file later. 2. The `copy_label()` function to library/functions/segmentation.
Seems like a fairly straingtforward refactor. See the arguments getting put together right before the library function call.
Finally, there's a TODO block - see the issue for details
