+++ src/subpackages/library/cellprofiler_library/functions/segmentation.py
@@ -849,3 +849,19 @@ def histogram_from_ijv(
     """
     return find_ijv_overlaps(parent_ijv, child_ijv, validate=True)
 
+###############################################################################
+# SplitOrMergeObjects
+###############################################################################
+
+def copy_labels(labels, segmented):
+    """Carry differences between orig_segmented and new_segmented into "labels"
+
+    labels - labels matrix similarly segmented to "segmented"
+    segmented - the newly numbered labels matrix (a subset of pixels are labeled)
+    """
+    max_labels = len(np.unique(segmented))
+    seglabel = scipy.ndimage.minimum(labels, segmented, np.arange(1, max_labels + 1))
+    labels_new = labels.copy()
+    labels_new[segmented != 0] = seglabel[segmented[segmented != 0] - 1]
+    return labels_new
+
